refactor: move CORS, Logger, Recovery, and RequestID middleware to its own file
