--- internal/infrastructure/middleware/global.go
@@ -1,80 +1,0 @@
-package middleware
-
-import (
-	"time"
-
-	"github.com/FIAP-SOAT-G20/hackathon-video-service/internal/infrastructure/logger"
-	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
-)
-
-func Logger(log *logger.Logger) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		start := time.Now()
-		path := c.Request.URL.Path
-		raw := c.Request.URL.RawQuery
-		requestID := c.GetString("request_id")
-
-		c.Next()
-
-		// Check if response came from cache
-		_, exists := c.Get("cache_miss")
-
-		log.InfoContext(c.Request.Context(),
-			"request completed",
-			"method", c.Request.Method,
-			"path", path,
-			"raw_query", raw,
-			"status", c.Writer.Status(),
-			"latency", time.Since(start),
-			"client_ip", c.ClientIP(),
-			"request_id", requestID,
-			"from_cache", !exists,
-			"errors", c.Errors.String(),
-		)
-	}
-}
-
-func RequestID() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		requestID := c.GetHeader("X-Request-ID")
-		if requestID == "" {
-			requestID = uuid.New().String()
-		}
-		c.Set("request_id", requestID)
-		c.Header("X-Request-ID", requestID)
-		c.Next()
-	}
-}
-
-func CORS() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
-		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
-		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
-		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE")
-
-		if c.Request.Method == "OPTIONS" {
-			c.AbortWithStatus(204)
-			return
-		}
-
-		c.Next()
-	}
-}
-
-func Recovery(log *logger.Logger) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		defer func() {
-			if err := recover(); err != nil {
-				log.ErrorContext(c.Request.Context(),
-					"panic recovered",
-					"error", err,
-					"request_id", c.GetString("request_id"),
-				)
-				c.AbortWithStatus(500)
-			}
-		}()
-		c.Next()
-	}
-}
